ducktape-deps: add packer cleanup script
This script runs only in the packer flavor of the image generation and
removes locations that use a lot of space but which are not needed in
the final image.

In the Docker version of image generation these are
generally excluded because they are on a "non mainline" part of the
multi-root build (which only have specific outputs copied to the final
image, discarding intermediates), or because they are captured by a
cache mount which also excludes them from the final image.

The packer build has no "non mainline" part so everything left behind
ends up in the final image, which is more than twice as large as the
Docker image.

This script helps to narrow that gap.